+++ src/main/java/com/nagarro/requests/NoteRequest.java
@@ -1,35 +1,35 @@
 package com.nagarro.requests;
 
 public class NoteRequest {
-	 private Integer id;
-	 private Integer task_id;
-	 private Integer list_id;
+	 private Long id;
+	 private Long task_id;
+	 private Long list_id;
 	 private String content;
 	 private String created_at;
 	 private String updated_at;
-	 private Integer revision;
+	 private Long revision;
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
-	public Integer getTask_id() {
+	public Long getTask_id() {
 		return task_id;
 	}
 
-	public void setTask_id(Integer task_id) {
+	public void setTask_id(Long task_id) {
 		this.task_id = task_id;
 	}
 
-	public Integer getList_id() {
+	public Long getList_id() {
 		return list_id;
 	}
 
-	public void setList_id(Integer list_id) {
+	public void setList_id(Long list_id) {
 		this.list_id = list_id;
 	}
 
@@ -57,11 +57,11 @@ public class NoteRequest {
 		this.updated_at = updated_at;
 	}
 
-	public Integer getRevision() {
+	public Long getRevision() {
 		return revision;
 	}
 
-	public void setRevision(Integer revision) {
+	public void setRevision(Long revision) {
 		this.revision = revision;
 	}
 }
